[JT-28] add tests for read-only-details screen
